+++ product/agent/daemon/src/test/java/se/crisp/codekvast/agent/daemon/codebase/scannertest/excluded/ExcludedScannerTest5.java
@@ -4,4 +4,9 @@ package se.crisp.codekvast.agent.daemon.codebase.scannertest.excluded;
  * @author Olle Hallin (qolha), olle.hallin@crisp.se
  */
 public class ExcludedScannerTest5 {
+
+    @SuppressWarnings("unused")
+    public void excludedSignature1() {
+
+    }
 }