+++ sample/sample-lib/src/main/java/sample/lib/excluded/ExcludedClass.java
@@ -1,0 +1,11 @@
+package sample.lib.excluded;
+
+/**
+ * @author olle.hallin@crisp.se
+ */
+@SuppressWarnings("unused")
+public class ExcludedClass {
+    public void foobar() {
+
+    }
+}